requirements.txt
dj-database-url==0.4.2
Django==1.10.5
gunicorn==19.6.0
psycopg2==2.6.2
whitenoise==3.2.3
djangorestframework==3.5.3
